Statement of goals and features
I made a first jet-lagged attempt to iron out the goals of the project, including the necessary and optional features. This can be extended to become the project's documentation. I think I followed to a large extent what was discussed in the last meeting. You can now categorize your user stories as part of one of the 4 "main features" (page types), or maybe some user stories may apply to more than one main feature. You can assess how many resources will be required for each of the main features and to what extent the optional features are achievable.